+++ src/Component/Data/ComponentData.php
@@ -2,8 +2,6 @@
 
 namespace Torr\Storyblok\Component\Data;
 
-use Torr\Storyblok\Component\AbstractComponent;
-
 /**
  * Wrapper around transformed component data.
  */
@@ -15,6 +13,5 @@ final class ComponentData
 		public readonly string $type,
 		public readonly array $data,
 		public readonly mixed $previewData = null,
-		public readonly AbstractComponent $component,
 	) {}
 }